api/pyproject.toml
# SentiBridge API
# Tier-based REST API for sentiment data access

[project]
name = "sentibridge-api"
version = "0.1.0"
description = "Tier-based REST API for SentiBridge sentiment oracle"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SentiBridge Team"}
]

dependencies = [
    # Web framework
    "fastapi>=0.108.0",
    "uvicorn[standard]>=0.25.0",
    "gunicorn>=21.2.0",
    
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    
    # Database
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.0",
    "redis>=5.0.0",
    "alembic>=1.13.0",
    
    # Web3
    "web3>=6.13.0",
    
    # Utilities
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.0",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
    
    # Rate limiting
    "slowapi>=0.1.8",
    
    # Monitoring
    "prometheus-fastapi-instrumentator>=6.1.0",
    "sentry-sdk[fastapi]>=1.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.25.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "black>=23.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E", "W", "F", "I", "C4", "B", "UP", "S", "A", "T20", "SIM", "RUF"
]
ignore = ["E501", "B008", "S101"]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101", "S106"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["web3.*", "slowapi.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
